Move PubKeyAndSignature into V1
